Add support for setting a coverage proportion
The idea is, instead of just needing to see this in two samples, we need to additionally have seen at least 80% of the genome (across all samples together).

This is implemented as simply as possible, by taking random start positions for reads and then just maintaining a set of the observed positions.  It is very slow, but not unusably slow.